ai-service/requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
requests==2.32.3
pandas==2.2.3
numpy<2.0.0
scikit-learn==1.5.2
torch==2.5.1
matplotlib==3.9.3
seaborn==0.13.2
neo4j>=5.0
